stuff is working

+++ scripts/ci/deploy.sh
@@ -10,17 +10,22 @@ export root_project_dir="$(pwd)"
 trap "cd \"$root_project_dir\"" EXIT
 
 
-rm -rf /tmp/deploy-skam-over-ssh
-mkdir /tmp/deploy-skam-over-ssh
-cd /tmp/deploy-skam-over-ssh
-git clone https://github.com/skamensky/skam.dev --depth 1
-cd skam.dev
-cp "$root_project_dir/scripts/env/secrets.env" ./scripts/env/secrets.env
+# useful when dev work required deployment and we don't want
+if [ "$1" == "--from-local" ]; then
+  echo "deploying from local"
+else
+  rm -rf /tmp/deploy-skam-over-ssh
+  mkdir /tmp/deploy-skam-over-ssh
+  cd /tmp/deploy-skam-over-ssh
+  git clone https://github.com/skamensky/skam.dev --depth 1
+  cd skam.dev
+  cp "$root_project_dir/scripts/env/secrets.env" ./scripts/env/secrets.env
+fi
 
 
 ssh_key_full_path="$HOME/$SSH_KEY_LOCATION_FROM_HOME"
-# replace these lines with your ssh credentials. Make sure docker and docker-compose are installed on the remote server
-rsync -Pav --exclude='.git/' --delete --delete-excluded -e "ssh -i $ssh_key_full_path" . ubuntu@$SSH_IP:/home/ubuntu/skam.dev
+# thanks to https://stackoverflow.com/a/15373763/4188138 from the ":- .gitignore" trick
+rsync -Pav --filter=':- .gitignore' --exclude='.git/' --delete --delete-excluded -e "ssh -i $ssh_key_full_path" . ubuntu@$SSH_IP:/home/ubuntu/skam.dev
+# copy the secrets file over
+scp -i "$ssh_key_full_path" "scripts/env/secrets.env" ubuntu@$SSH_IP:/home/ubuntu/skam.dev/scripts/env/secrets.env
 ssh -i $ssh_key_full_path ubuntu@$SSH_IP -t "export STAGE=prod; cd skam.dev && /bin/bash ./scripts/docker/ops.sh up"
-
-cd $root_project_dir
